lesson nr 3 description and reformatted code

+++ Lecture03/inversionCount.cpp
@@ -5,13 +5,13 @@
 #include <climits>
 
 
-std::pair<std::vector<uint64_t>, long long>
-mergeCount(std::pair<std::vector<uint64_t>, long long> left, std::pair<std::vector<uint64_t>, long long> right) {
-    std::pair<std::vector<uint64_t>, long long> result(std::vector<uint64_t>(left.first.size() + right.first.size()),
-                                                 left.second + right.second);
-    long long i = 0;
-    long long j = 0;
-    long long aus = 0;
+std::pair<std::vector<uint64_t>, uint64_t>
+mergeCount(std::pair<std::vector<uint64_t>, uint64_t> left, std::pair<std::vector<uint64_t>, uint64_t> right) {
+    std::pair<std::vector<uint64_t>, uint64_t> result(std::vector<uint64_t>(left.first.size() + right.first.size()),
+                                                      left.second + right.second);
+    uint64_t i = 0;
+    uint64_t j = 0;
+    uint64_t aus = 0;
     left.first.push_back(INT_MAX);
     right.first.push_back(INT_MAX);
     while (i < left.first.size() - 1 || j < right.first.size() - 1) {
@@ -28,15 +28,15 @@ mergeCount(std::pair<std::vector<uint64_t>, long long> left, std::pair<std::vect
     return result;
 }
 
-std::pair<std::vector<uint64_t>, long long> inversionCount(std::vector<uint64_t> &v, int start, int end) {
+std::pair<std::vector<uint64_t>, uint64_t> inversionCount(std::vector<uint64_t> &v, int start, int end) {
     if (start >= end) {
         std::vector<uint64_t> base(1);
         base[0] = v[end];
-        return std::pair<std::vector<uint64_t>, long long>(base, 0);
+        return std::pair<std::vector<uint64_t>, uint64_t>(base, 0);
     } else {
         int mid = start + ((end - start) / 2);
-        std::pair<std::vector<uint64_t>, long long> resLeft = inversionCount(v, start, mid);
-        std::pair<std::vector<uint64_t>, long long> resRight = inversionCount(v, mid + 1, end);
+        std::pair<std::vector<uint64_t>, uint64_t> resLeft = inversionCount(v, start, mid);
+        std::pair<std::vector<uint64_t>, uint64_t> resRight = inversionCount(v, mid + 1, end);
         return mergeCount(resLeft, resRight);
     }
 
@@ -49,17 +49,17 @@ int main() {
     int t;
     std::cin >> t;
     for (int i = 0; i < t; ++i) {
-        long long n;
+        uint64_t n;
         std::cin >> n;
         std::vector<uint64_t> v;
         v.reserve(n);
-        for (long long j = 0; j < n; ++j) {
+        for (uint64_t j = 0; j < n; ++j) {
             uint64_t in;
             std::cin >> in;
             v.push_back(in);
         }
 
-        std::pair<std::vector<uint64_t>, long long> result = inversionCount(v, 0, v.size() - 1);
+        std::pair<std::vector<uint64_t>, uint64_t> result = inversionCount(v, 0, v.size() - 1);
         std::cout << result.second << std::endl;
         v.clear();
     }
